refactor(user-profile): ユーザーメニューと QR モーダルを分離 (#707)
UserProfileContent.vue (3140 行) の god component 分割 第 1 段。
凝集した 2 つの塊を window/user-profile/ 配下へ切り出す:

- UserProfileMenu: PopupMenu 一式 (ミュート/ブロック/リノートミュート/
  フォロワー解除/通報/コピー系/リスト・アンテナ追加/フォロー設定/検索/DM)。
  userRelation の取得・保持もメニューだけが使うためここへ移動。
  投稿フォームは親所有のため「ユーザー指定ノートを作成」は compose emit
- UserProfileQrCode: QR コードモーダル (qr-code-styling の遅延 import ごと)

adapter は子へ渡すため let → shallowRef に変更 (代入は onMounted の 1 回)。
3140 行 → 2386 行。見た目・挙動は変更なし。
typecheck / vitest 1258 件 / vite build で確認。
